requirements.txt
streamlit==1.39.0
torch==2.1.0
transformers==4.36.0
google-generativeai==0.3.2
moviepy==1.0.3
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.4
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.31.0

